app-previsional-ia-main/requirements.txt
streamlit
google-generativeai
python-docx
PyMuPDF
pytesseract
Pillow
docxtpl
